Refactor package names across modules for consistency
