+++ modules/http/src/main/scala/ch/epfl/bluebrain/nexus/commons/http/UnexpectedUnsuccessfulHttpResponse.scala
@@ -1,7 +1,7 @@
 package ch.epfl.bluebrain.nexus.commons.http
 
 import akka.http.scaladsl.model.HttpResponse
-import ch.epfl.bluebrain.nexus.common.types.Err
+import ch.epfl.bluebrain.nexus.commons.types.Err
 
 /**
   * Error type representing an unexpected unsuccessful http response.  Its entity bytes are discarded.
