+++ modules/service-commons/src/main/scala/ch/epfl/bluebrain/nexus/commons/service/io/package.scala
@@ -1,0 +1,12 @@
+package ch.epfl.bluebrain.nexus.commons.service
+
+import java.nio.charset.Charset
+
+package object io {
+
+  /**
+    * Constant value for the UTF-8 charset.
+    */
+  final val UTF8: Charset = Charset.forName("UTF-8")
+
+}
