--- modules/service-commons/src/main/scala/ch/epfl/bluebrain/nexus/service/commons/io/package.scala
@@ -1,12 +1,0 @@
-package ch.epfl.bluebrain.nexus.service.commons
-
-import java.nio.charset.Charset
-
-package object io {
-
-  /**
-    * Constant value for the UTF-8 charset.
-    */
-  final val UTF8: Charset = Charset.forName("UTF-8")
-
-}
