+++ modules/service-commons/src/test/resources/cassandra.conf
@@ -2,10 +2,10 @@ cassandra-journal {
   contact-points = ["127.0.0.1:"${test.cassandra-port}]
 
   event-adapters {
-    tagging = "ch.epfl.bluebrain.nexus.service.commons.persistence.Fixture$TaggingAdapter"
+    tagging = "ch.epfl.bluebrain.nexus.commons.service.persistence.Fixture$TaggingAdapter"
   }
   event-adapter-bindings {
-    "ch.epfl.bluebrain.nexus.service.commons.persistence.Fixture$Event" = tagging
+    "ch.epfl.bluebrain.nexus.commons.service.persistence.Fixture$Event" = tagging
   }
 }
